% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.R
\docType{data}
\name{geno}
\alias{geno}
\title{somedata to test}
\format{
A matrix containing genotype information
}
\usage{
data(geno)
}
\description{
Numeric format of genotype dataset
}
\keyword{datasets}
\keyword{genotype}
