% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRV1.R
\name{SEIR}
\alias{SEIR}
\title{Selector-Embedded Iterative Regression}
\usage{
SEIR(
  Y,
  X,
  GM = NULL,
  CV = NULL,
  maxStep = 10,
  selector = c("L0", "L1", "MCP", "stepwise", "bess", "MARS", "SCAD", "lars", "vif",
    "SES", "FBED"),
  fun.selection = c("fastLm", "lm.vec"),
  extraction = c("bagging-like", "1se", "min", "mean"),
  X.check = c(TRUE, FALSE),
  chunk.num = NULL,
  file.output = TRUE,
  plot.style = "CMplot",
  cutOff = 0.05
)
}
\arguments{
\item{Y}{Phenotypic vector with length n}

\item{X}{n by m dataframe, matrix or big.matrix of genotype matrix}

\item{GM}{m by 3 dataframe or matrix for SNP name, chromosome and BP}

\item{CV}{n by t dataframe or matrix for t covariates}

\item{maxStep}{maxStep Maximum number of iteration steps}

\item{selector}{selector A method for Variable selection}

\item{fun.selection}{fun.selection The core function for parameter calculation, lm.vec is more faster than fastLm}

\item{extraction}{A method for calculating p values}

\item{X.check}{logical}

\item{chunk.num}{The "pieces" number of genotype matrix when considering memory size}

\item{file.output}{output GWAS results, Manhattan and QQ plot}

\item{plot.style}{plot.style}

\item{cutOff}{The threshold}
}
\value{
List - includes the suggested SNPs, the results with normal 4-column format (SNP, chromosome, BP, p-values), the matrix of coefficients, SE, z-values, p-values, and the matrices of coefficients, SE, z-values, p-values from each iterative step
}
\description{
Selector-Embedded Iterative Regression
}
\examples{
         data(geno)
         data(phe)
         data(map)
         SEIR(Y=phe,X=geno,GM=map,CV=NULL,maxStep=10,
              selector="stepwise",fun.selection="fastLm",
              extraction="min",X.check="FALSE",chunk.num = NULL,
              file.output=FALSE,plot.style="SEIR",cutOff=0.05)
}
